+++ cypress/integration/custom_extension_variable_type.ts
@@ -1,3 +1,5 @@
+import * as constants from "../../projects/ng-rule-editor/src/lib/validation";
+
 describe('Rule editor', () => {
   beforeEach(() => {
     cy.visit('/');
@@ -222,7 +224,7 @@ describe('Rule editor', () => {
           // The FHIRPath expression is now being validated using fhirpath.js. So this will
           // result in the "Invalid expression." error.
           cy.get('#variable-expression-19').should('have.class', 'field-error');
-          cy.get('#expression-error').should('contain.text', 'Invalid expression.');
+          cy.get('#expression-error').should('contain.text', constants.INVALID_EXPRESSION);
         });
 
         // Variable type "Question" is displayed correctly
@@ -508,7 +510,7 @@ describe('Rule editor', () => {
 
         // There is an error in the Output Expression
         cy.get('#simple-expression-final').should('have.class', 'field-error');
-        cy.get('#final-expression-section #expression-error > p').should('contain.text', 'Invalid expression.');
+        cy.get('#final-expression-section #expression-error > p').should('contain.text', constants.INVALID_EXPRESSION);
 
         // As a result, the Save button is disabled
         cy.get('#export').should('have.class', 'disabled');
@@ -541,23 +543,23 @@ describe('Rule editor', () => {
         // should throw "Expression is required." b/c they are empty.
         cy.get('div#row-0').within(() => {
           cy.get('#simple-expression-0').should('have.class', 'field-error');
-          cy.get('#expression-error > p').should('contain.text', 'Expression is required.');
+          cy.get('#expression-error > p').should('contain.text', constants.EXPRESSION_REQUIRED);
         });
         cy.get('div#row-8').within(() => {
           cy.get('#autocomplete-8').should('have.class', 'field-error');
-          cy.get('#expression-error > p').should('contain.text', 'Expression is required.');
+          cy.get('#expression-error > p').should('contain.text', constants.FHIR_QUERY_OBSERVATION_REQUIRED);
         });
         cy.get('div#row-10').within(() => {
           cy.get('#question-10').should('have.class', 'field-error');
-          cy.get('#expression-error > p').should('contain.text', 'Expression is required.');
+          cy.get('#expression-error > p').should('contain.text', constants.QUESTION_REQUIRED);
         });
         cy.get('div#row-23').within(() => {
           cy.get('#question-23').should('have.class', 'field-error');
-          cy.get('#expression-error > p').should('contain.text', 'Expression is required.');
+          cy.get('#expression-error > p').should('contain.text', constants.QUESTION_REQUIRED);
         });
         cy.get('div#row-25').within(() => {
           cy.get('#variable-expression-25').should('have.class', 'field-error');
-          cy.get('#expression-error > p').should('contain.text', 'Expression is required.');
+          cy.get('#expression-error > p').should('contain.text', constants.EXPRESSION_REQUIRED);
         });
 
         // Fix the errors by entering data