+++ cypress/integration/spec.ts
@@ -1,3 +1,5 @@
+import * as constants from "../../projects/ng-rule-editor/src/lib/validation";
+
 describe('Rule editor', () => {
   beforeEach(() => {
     cy.visit('/');
@@ -184,7 +186,7 @@ describe('Rule editor', () => {
 
         // There is an error in the Output Expression
         cy.get('#simple-expression-final').should('have.class', 'field-error');
-        cy.get('#final-expression-section #expression-error > p').should('contain.text', 'Expression is required.');
+        cy.get('#final-expression-section #expression-error > p').should('contain.text', constants.EXPRESSION_REQUIRED);
         // As a result, the Save button is disabled
         cy.get('#export').should('have.class', 'disabled');
         
@@ -1193,7 +1195,7 @@ describe('Rule editor', () => {
         cy.get('#variables-section .variable-row').should('have.length', 2);
   
         // Confirm that variable b is no longer available for Output expression
-        cy.get('#final-expression-section #expression-error > p').should('contain.text', 'Invalid expression.');
+        cy.get('#final-expression-section #expression-error > p').should('contain.text', constants.INVALID_EXPRESSION);
 
         // Confirm that expression without variable b is valid
         cy.get('#simple-expression-final').clear().type('a + c');