+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project follows [Semantic Versioning](http://semver.org/).
 ### Added
 - Added expression validation
 - Implemented logic to disable the 'Save' button if validation fails
+- Updated autocomplete fields to accept blank as an input
 
 ## [3.1.16] 2024-01-17
 ### Changed